+++ src/serializers/type_serializers/any.rs
@@ -1,20 +1,17 @@
 use std::borrow::Cow;
 use std::str::from_utf8;
 
-use pyo3::ffi::PyTypeObject;
-use pyo3::once_cell::GILOnceCell;
 use pyo3::prelude::*;
 use pyo3::types::{
     PyByteArray, PyBytes, PyDate, PyDateTime, PyDelta, PyDict, PyFrozenSet, PyList, PySet, PyString, PyTime, PyTuple,
 };
 
 use serde::ser::{Serialize, SerializeMap, SerializeSeq, Serializer};
-use strum_macros::EnumString;
 
 use crate::build_context::BuildContext;
 use crate::url::{PyMultiHostUrl, PyUrl};
 
-use super::shared::{py_err_se_err, BuildSerializer, CombinedSerializer, Extra, SerMode, TypeSerializer};
+use super::{py_err_se_err, BuildSerializer, CombinedSerializer, Extra, ObType, SerMode, TypeSerializer};
 
 #[derive(Debug, Clone)]
 pub struct AnySerializer;
@@ -44,18 +41,18 @@ impl TypeSerializer for AnySerializer {
     }
 }
 
-pub(super) fn fallback_to_python(value: &PyAny, extra: &Extra) -> PyResult<PyObject> {
+pub(crate) fn fallback_to_python(value: &PyAny, extra: &Extra) -> PyResult<PyObject> {
     match extra.mode {
         SerMode::Json => fallback_to_python_json(value, extra),
         _ => Ok(value.into_py(value.py())),
     }
 }
 
-pub(super) fn fallback_to_python_json(value: &PyAny, extra: &Extra) -> PyResult<PyObject> {
+pub(crate) fn fallback_to_python_json(value: &PyAny, extra: &Extra) -> PyResult<PyObject> {
     ob_type_to_python_json(&extra.ob_type_lookup.get_type(value), value, extra)
 }
 
-pub(super) fn ob_type_to_python_json(ob_type: &ObType, value: &PyAny, extra: &Extra) -> PyResult<PyObject> {
+pub(crate) fn ob_type_to_python_json(ob_type: &ObType, value: &PyAny, extra: &Extra) -> PyResult<PyObject> {
     let value_id = extra.rec_guard.add(value)?;
     let py = value.py();
 
@@ -143,13 +140,13 @@ pub(super) fn ob_type_to_python_json(ob_type: &ObType, value: &PyAny, extra: &Ex
     Ok(value)
 }
 
-pub(super) struct SerializeInfer<'py> {
+pub(crate) struct SerializeInfer<'py> {
     value: &'py PyAny,
     extra: &'py Extra<'py>,
 }
 
 impl<'py> SerializeInfer<'py> {
-    pub(super) fn new(value: &'py PyAny, extra: &'py Extra) -> Self {
+    pub(crate) fn new(value: &'py PyAny, extra: &'py Extra) -> Self {
         Self { value, extra }
     }
 }
@@ -161,7 +158,7 @@ impl<'py> Serialize for SerializeInfer<'py> {
     }
 }
 
-pub(super) fn fallback_serialize<S: Serializer>(
+pub(crate) fn fallback_serialize<S: Serializer>(
     value: &PyAny,
     serializer: S,
     extra: &Extra,
@@ -169,7 +166,7 @@ pub(super) fn fallback_serialize<S: Serializer>(
     fallback_serialize_known(&extra.ob_type_lookup.get_type(value), value, serializer, extra)
 }
 
-pub(super) fn fallback_serialize_known<S: Serializer>(
+pub(crate) fn fallback_serialize_known<S: Serializer>(
     ob_type: &ObType,
     value: &PyAny,
     serializer: S,
@@ -266,7 +263,7 @@ pub(super) fn fallback_serialize_known<S: Serializer>(
     ser_result
 }
 
-pub(super) fn json_key<'py>(key: &'py PyAny, extra: &Extra) -> PyResult<Cow<'py, str>> {
+pub(crate) fn json_key<'py>(key: &'py PyAny, extra: &Extra) -> PyResult<Cow<'py, str>> {
     let ob_type = extra.ob_type_lookup.get_type(key);
 
     match ob_type {
@@ -314,211 +311,4 @@ pub(super) fn json_key<'py>(key: &'py PyAny, extra: &Extra) -> PyResult<Cow<'py,
         }
         _ => Ok(key.str()?.to_string_lossy()),
     }
-}
-
-#[derive(Debug, Clone)]
-pub struct ObTypeLookup {
-    none: usize,
-    // numeric types
-    int: usize,
-    bool: usize,
-    float: usize,
-    // string types
-    string: usize,
-    bytes: usize,
-    bytearray: usize,
-    // mapping types
-    dict: usize,
-    // sequence types
-    list: usize,
-    tuple: usize,
-    set: usize,
-    frozenset: usize,
-    // datetime types
-    datetime: usize,
-    date: usize,
-    time: usize,
-    timedelta: usize,
-    // types from this package
-    url: usize,
-    multi_host_url: usize,
-}
-
-static TYPE_LOOKUP: GILOnceCell<ObTypeLookup> = GILOnceCell::new();
-
-pub enum IsType {
-    Exact,
-    Subclass,
-    False,
-}
-
-impl ObTypeLookup {
-    fn new(py: Python) -> Self {
-        let lib_url = url::Url::parse("https://example.com").unwrap();
-        Self {
-            none: py.None().as_ref(py).get_type_ptr() as usize,
-            // numeric types
-            int: 0i32.into_py(py).as_ref(py).get_type_ptr() as usize,
-            bool: true.into_py(py).as_ref(py).get_type_ptr() as usize,
-            float: 0f32.into_py(py).as_ref(py).get_type_ptr() as usize,
-            // string types
-            string: PyString::new(py, "s").get_type_ptr() as usize,
-            bytes: PyBytes::new(py, b"s").get_type_ptr() as usize,
-            bytearray: PyByteArray::new(py, b"s").get_type_ptr() as usize,
-            // sequence types
-            list: PyList::empty(py).get_type_ptr() as usize,
-            tuple: PyTuple::empty(py).get_type_ptr() as usize,
-            set: PySet::empty(py).unwrap().get_type_ptr() as usize,
-            frozenset: PyFrozenSet::empty(py).unwrap().get_type_ptr() as usize,
-            // mapping types
-            dict: PyDict::new(py).get_type_ptr() as usize,
-            // datetime types
-            datetime: PyDateTime::new(py, 2000, 1, 1, 0, 0, 0, 0, None)
-                .unwrap()
-                .get_type_ptr() as usize,
-            date: PyDate::new(py, 2000, 1, 1).unwrap().get_type_ptr() as usize,
-            time: PyTime::new(py, 0, 0, 0, 0, None).unwrap().get_type_ptr() as usize,
-            timedelta: PyDelta::new(py, 0, 0, 0, false).unwrap().get_type_ptr() as usize,
-            // types from this package
-            url: PyUrl::new(lib_url.clone()).into_py(py).as_ref(py).get_type_ptr() as usize,
-            multi_host_url: PyMultiHostUrl::new(lib_url, None).into_py(py).as_ref(py).get_type_ptr() as usize,
-        }
-    }
-
-    pub fn cached(py: Python<'_>) -> &Self {
-        TYPE_LOOKUP.get_or_init(py, || Self::new(py))
-    }
-
-    pub fn is_type(&self, value: &PyAny, expected_ob_type: ObType) -> IsType {
-        self.ob_type_is_expected(value.get_type_ptr(), expected_ob_type)
-    }
-
-    fn ob_type_is_expected(&self, type_ptr: *mut PyTypeObject, expected_ob_type: ObType) -> IsType {
-        let ob_type = type_ptr as usize;
-        let ans = match expected_ob_type {
-            ObType::None => self.none == ob_type,
-            ObType::Int => self.int == ob_type,
-            ObType::Bool => self.bool == ob_type,
-            ObType::Float => self.float == ob_type,
-            ObType::Str => self.string == ob_type,
-            ObType::Dict => self.dict == ob_type,
-            ObType::List => self.list == ob_type,
-            ObType::Tuple => self.tuple == ob_type,
-            ObType::Set => self.set == ob_type,
-            ObType::Frozenset => self.frozenset == ob_type,
-            ObType::Bytes => self.bytes == ob_type,
-            ObType::Datetime => self.datetime == ob_type,
-            ObType::Date => self.date == ob_type,
-            ObType::Time => self.time == ob_type,
-            ObType::Timedelta => self.timedelta == ob_type,
-            ObType::Bytearray => self.bytearray == ob_type,
-            ObType::Url => self.url == ob_type,
-            ObType::MultiHostUrl => self.multi_host_url == ob_type,
-            ObType::Unknown => false,
-        };
-
-        if ans {
-            IsType::Exact
-        } else {
-            // this allows for subtypes of the supported class types,
-            // if we didn't successfully confirm the type, we try again with the next base type pointer provided
-            // it's not null
-            let base_type_ptr = unsafe { (*type_ptr).tp_base };
-            if base_type_ptr.is_null() {
-                IsType::False
-            } else {
-                match self.ob_type_is_expected(base_type_ptr, expected_ob_type) {
-                    IsType::False => IsType::False,
-                    _ => IsType::Subclass,
-                }
-            }
-        }
-    }
-
-    pub fn get_type(&self, value: &PyAny) -> ObType {
-        self.lookup_by_ob_type(value.get_type_ptr())
-    }
-
-    fn lookup_by_ob_type(&self, type_ptr: *mut PyTypeObject) -> ObType {
-        let ob_type = type_ptr as usize;
-        // this should be pretty fast, but still order is a bit important, so the most common types should come first
-        // thus we don't follow the order of ObType
-        if ob_type == self.none {
-            ObType::None
-        } else if ob_type == self.int {
-            ObType::Int
-        } else if ob_type == self.bool {
-            ObType::Bool
-        } else if ob_type == self.float {
-            ObType::Float
-        } else if ob_type == self.string {
-            ObType::Str
-        } else if ob_type == self.dict {
-            ObType::Dict
-        } else if ob_type == self.list {
-            ObType::List
-        } else if ob_type == self.tuple {
-            ObType::Tuple
-        } else if ob_type == self.set {
-            ObType::Set
-        } else if ob_type == self.frozenset {
-            ObType::Frozenset
-        } else if ob_type == self.bytes {
-            ObType::Bytes
-        } else if ob_type == self.datetime {
-            ObType::Datetime
-        } else if ob_type == self.date {
-            ObType::Date
-        } else if ob_type == self.time {
-            ObType::Time
-        } else if ob_type == self.timedelta {
-            ObType::Timedelta
-        } else if ob_type == self.bytearray {
-            ObType::Bytearray
-        } else if ob_type == self.url {
-            ObType::Url
-        } else if ob_type == self.multi_host_url {
-            ObType::MultiHostUrl
-        } else {
-            // this allows for subtypes of the supported class types,
-            // if `ob_type` didn't match any member of self, we try again with the next base type pointer
-            let base_type_ptr = unsafe { (*type_ptr).tp_base };
-            if base_type_ptr.is_null() {
-                ObType::Unknown
-            } else {
-                self.lookup_by_ob_type(base_type_ptr)
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone, EnumString)]
-#[strum(serialize_all = "snake_case")]
-pub enum ObType {
-    None,
-    // numeric types
-    Int,
-    Bool,
-    Float,
-    // string types
-    Str,
-    Bytes,
-    Bytearray,
-    // sequence types
-    List,
-    Tuple,
-    Set,
-    Frozenset,
-    // mapping types
-    Dict,
-    // datetime types
-    Datetime,
-    Date,
-    Time,
-    Timedelta,
-    // types from this package
-    Url,
-    MultiHostUrl,
-    // unknown type
-    Unknown,
 }
