+++ src/serializers/type_serializers/datetime_etc.rs
@@ -7,7 +7,7 @@ use crate::build_context::BuildContext;
 use crate::input::{pydate_as_date, pydatetime_as_datetime, pytime_as_time};
 
 use super::any::{fallback_serialize, fallback_to_python_json, json_key};
-use super::shared::{py_err_se_err, BuildSerializer, CombinedSerializer, Extra, SerMode, TypeSerializer};
+use super::{py_err_se_err, BuildSerializer, CombinedSerializer, Extra, SerMode, TypeSerializer};
 
 pub(crate) fn datetime_to_string(py_dt: &PyDateTime) -> PyResult<String> {
     let dt = pydatetime_as_datetime(py_dt)?;
