+++ src/serializers/type_serializers/tuple.rs
@@ -8,9 +8,8 @@ use crate::build_context::BuildContext;
 use crate::build_tools::SchemaDict;
 
 use super::any::{fallback_serialize, fallback_to_python, AnySerializer};
-use super::include_exclude::SchemaIncEx;
-use super::shared::{
-    py_err_se_err, BuildSerializer, CombinedSerializer, Extra, PydanticSerializer, SerMode, TypeSerializer,
+use super::{
+    py_err_se_err, BuildSerializer, CombinedSerializer, Extra, PydanticSerializer, SchemaIncEx, SerMode, TypeSerializer,
 };
 
 pub struct TupleBuilder;
