+++ src/serializers/type_serializers/with_default.rs
@@ -8,7 +8,7 @@ use crate::build_context::BuildContext;
 use crate::build_tools::SchemaDict;
 use crate::validators::DefaultType;
 
-use super::shared::{BuildSerializer, CombinedSerializer, Extra, TypeSerializer};
+use super::{BuildSerializer, CombinedSerializer, Extra, TypeSerializer};
 
 #[derive(Debug, Clone)]
 pub struct WithDefaultSerializer {
